refactor move get_metadata endpoint into dedicated module
- created python/mdvtools/metadata/ package
- extracted /get_metadata endpoint logic (lines 160–246 from server.py) into metadata/routes.py
- updated server.py to import and register routes via register_metadata_routes(project_bp, project)
- removed original endpoint code from server.py
- cleaned up unused imports (now scoped to the metadata module)
