+++ python/mdvtools/server.py
@@ -33,7 +33,7 @@ import scanpy as sc
 from mdvtools.conversions import convert_scanpy_to_mdv
 from mdvtools.server_extension import MDVServerOptions
 from mdvtools.logging_config import get_logger
-from mdvtools.zarr_utils import get_zarr_extractor
+from mdvtools.metadata import register_metadata_routes
 
 
 logger = get_logger(__name__)
@@ -103,6 +103,9 @@ def create_app(
     for extension in options.extensions:
         extension.register_routes(project, project_bp)
 
+    # Register metadata routes
+    register_metadata_routes(project_bp, project)
+
 
     @project_bp.route("/")
     def project_index():
@@ -156,94 +159,6 @@ def create_app(
                     return f"Problem parsing state file: {e}", 500
         return send_file(path)
 
-    # Universal metadata endpoint for fetching remote dataset information
-    @project_bp.route("/get_metadata", methods=["GET"])
-    def get_metadata():
-        """
-        Fetch metadata from a dataset URL (supports Zarr and SpatialData formats).
-        
-        Query Parameters:
-            url: The URL of the dataset to fetch metadata from
-            
-        Returns:
-            JSON response with metadata structure or error message
-        """
-        log("=== METADATA REQUEST ===")
-        log(f"Request method: {request.method}")
-        log(f"Request path: {request.path}")
-        log(f"Request args: {request.args}")
-        log(f"Project: {project.id}")
-        
-        dataset_url = None
-        try:
-            dataset_url = request.args.get('url')
-            log(f"Extracted dataset_url: {dataset_url}")
-            
-            if not dataset_url:
-                log("ERROR: Missing 'url' parameter")
-                return jsonify({
-                    "error": "Missing 'url' parameter",
-                    "message": "Please provide a 'url' query parameter with the dataset URL"
-                }), 400
-                
-            # Validate URL format
-            if not dataset_url.startswith(('http://', 'https://')):
-                log(f"ERROR: Invalid URL format: {dataset_url}")
-                return jsonify({
-                    "error": "Invalid URL format", 
-                    "message": "URL must start with http:// or https://"
-                }), 400
-            
-            # Convert localhost to host.docker.internal if running in Docker
-            if 'localhost' in dataset_url and os.path.exists('/.dockerenv'):
-                dataset_url = dataset_url.replace('localhost', 'host.docker.internal')
-                log(f"Running in Docker, converted URL to: {dataset_url}")
-            
-            log(f"URL validation passed, fetching metadata for: {dataset_url}")
-            
-            # Extract metadata using zarr utilities (supports both Zarr and SpatialData)
-            log("Getting metadata extractor...")
-            extractor = get_zarr_extractor()
-            log(f"Metadata extractor: {extractor}")
-            
-            log("Calling fetch_zarr_metadata...")
-            print("Creating new event loop for async metadata call...")
-            # Run the async function in the current event loop
-            loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(loop)
-            try:
-                print("Running fetch_zarr_metadata in event loop...")
-                metadata = loop.run_until_complete(extractor.fetch_zarr_metadata(dataset_url))
-                print(f"Event loop completed, metadata type: {type(metadata)}")
-            except Exception as loop_e:
-                print(f"Exception in event loop: {loop_e}")
-                raise loop_e
-            finally:
-                print("Closing event loop...")
-                loop.close()
-                print("Event loop closed")
-            log(f"Metadata extraction successful, keys: {list(metadata.keys()) if isinstance(metadata, dict) else type(metadata)}")
-            print(f"Metadata extraction successful, keys: {list(metadata.keys()) if isinstance(metadata, dict) else type(metadata)}")
-            
-            response_data = {
-                "success": True,
-                "metadata": metadata,
-                "url": dataset_url
-            }
-            log("Returning successful response")
-            return jsonify(response_data)
-            
-        except Exception as e:
-            log(f"EXCEPTION in get_metadata: {str(e)}")
-            log(f"Exception type: {type(e)}")
-            import traceback
-            log(f"Full traceback: {traceback.format_exc()}")
-            
-            return jsonify({
-                "error": "Metadata extraction failed",
-                "message": str(e),
-                "url": dataset_url
-            }), 500
 
     # gets the raw byte data and packages it in the correct response
     @project_bp.route("/get_data", methods=["POST"])
